Made Lua more functional
Basic functionality added to lua support, increased version to 0.2 to
mark the usability of lua

+++ source/world/terraingen/terraingen.cpp
@@ -25,7 +25,7 @@ void CTerrainGenerator::destroy()
 
 CBlock* CTerrainGenerator::generateBlock( glm::ivec2 chunkPos, glm::ivec2 blockPos )
 {
-	int grassLevel = (m_perlin.octaveNoise( (float)chunkPos.x / 16.0f, 2 ) * 64) + GrassDepth;
+	int grassLevel = (int)((m_perlin.octaveNoise( (float)chunkPos.x / 16.0f, 2 ) * 64) + GrassDepth);
 
 	if( chunkPos.y == SurfaceLevel ) {
 		if( blockPos.y <= grassLevel ) {
